+++ led-effects/src/effects/modifiers/temporal.rs
@@ -1,5 +1,3 @@
-// fade
-
 use crate::effects::core::{
     pixel::Pixel,
     traits::{Modifier, Parameter},
@@ -7,17 +5,17 @@ use crate::effects::core::{
 
 pub struct Decay<Rate>
 where
-    Rate: Parameter<f32>,
+    Rate: Parameter<u8>,
 {
-    pub rate: Rate,
+    pub rate: Rate, // 0-255 (where 255 = no decay, 0 = instant fade)
 }
 
 impl<Rate> Modifier for Decay<Rate>
 where
-    Rate: Parameter<f32>,
+    Rate: Parameter<u8>,
 {
     #[inline(always)]
-    fn modify(&mut self, buffer: &mut [Pixel], now: u64) {
+    fn modify(&mut self, buffer: &mut [Pixel], now: u32) {
         let rate = self.rate.sample(now);
         for pixel in buffer.iter_mut() {
             *pixel = pixel.scale(rate);
